README: provide corresponding source for old commits
d718bef (certdata.txt: the corresponding source, 2024-10-29) made it
easy to find the corresponding source code for the current version of
ca-bundle.crt and for any future versions of ca-bundle.crt.

This commit makes it easy to find the corresponding source code for
previous versions of ca-bundle.crt.

In order to help verify that the information added by this commit is
accurate, I created this script [1]. That script will go through every
commit that comes before d718bef (certdata.txt: the corresponding
source, 2024-10-29) and try to build that commit’s version of
ca-bundle.crt using the corresponding source code that the README says
that that commit uses. It will also compare the newly built
ca-bundle.crt file to the one that’s in the old commit in order to make
sure that they’re identical.

Ref: #11

[1]: <https://gist.github.com/Jayman2000/53b7303775af9cc8a99537260d341ae3>